add-product/requirements.txt
oci
requests
parliament-functions==0.1.0